trying to organize the script in Metric class
